+++ concur.js
@@ -1,41 +1,62 @@
 /**
- * Concur 0.2.5 - https://github.com/insin/concur
+ * Concur 0.3.0 - https://github.com/insin/concur
  * MIT Licensed
  */
-!function(e){if("object"==typeof exports)module.exports=e();else if("function"==typeof define&&define.amd)define(e);else{var f;"undefined"!=typeof window?f=window:"undefined"!=typeof global?f=global:"undefined"!=typeof self&&(f=self),f.Concur=e()}}(function(){var define,module,exports;return (function e(t,n,r){function s(o,u){if(!n[o]){if(!t[o]){var a=typeof require=="function"&&require;if(!u&&a)return a(o,!0);if(i)return i(o,!0);throw new Error("Cannot find module '"+o+"'")}var f=n[o]={exports:{}};t[o][0].call(f.exports,function(e){var n=t[o][1][e];return s(n?n:e)},f,f.exports,e,t,n,r)}return n[o].exports}var i=typeof require=="function"&&require;for(var o=0;o<r.length;o++)s(r[o]);return s})({1:[function(require,module,exports){
+!function(e){if("object"==typeof exports&&"undefined"!=typeof module)module.exports=e();else if("function"==typeof define&&define.amd)define([],e);else{var f;"undefined"!=typeof window?f=window:"undefined"!=typeof global?f=global:"undefined"!=typeof self&&(f=self),f.Concur=e()}}(function(){var define,module,exports;return (function e(t,n,r){function s(o,u){if(!n[o]){if(!t[o]){var a=typeof require=="function"&&require;if(!u&&a)return a(o,!0);if(i)return i(o,!0);var f=new Error("Cannot find module '"+o+"'");throw f.code="MODULE_NOT_FOUND",f}var l=n[o]={exports:{}};t[o][0].call(l.exports,function(e){var n=t[o][1][e];return s(n?n:e)},l,l.exports,e,t,n,r)}return n[o].exports}var i=typeof require=="function"&&require;for(var o=0;o<r.length;o++)s(r[o]);return s})({1:[function(require,module,exports){
 'use strict';
 
-var is = require('isomorph/is')
-var object = require('isomorph/object')
+var hasOwn = Object.prototype.hasOwnProperty
+var toString = Object.prototype.toString
+
+function type(obj) {
+  return toString.call(obj).slice(8, -1).toLowerCase()
+}
+
+function inherits(childConstructor, parentConstructor) {
+  var F = function() {}
+  F.prototype = parentConstructor.prototype
+  childConstructor.prototype = new F()
+  childConstructor.prototype.constructor = childConstructor
+  return childConstructor
+}
+
+function extend(dest, src) {
+  for (var prop in src) {
+    if (hasOwn.call(src, prop)) {
+      dest[prop] = src[prop]
+    }
+  }
+  return dest
+}
 
 /**
  * Mixes in properties from one object to another. If the source object is a
  * Function, its prototype is mixed in instead.
  */
 function mixin(dest, src) {
-  if (is.Function(src)) {
-    object.extend(dest, src.prototype)
+  if (type(src) == 'function') {
+    extend(dest, src.prototype)
   }
   else {
-    object.extend(dest, src)
+    extend(dest, src)
   }
 }
 
 /**
- * Applies mixins specified as a __mixin__ property on the given properties
+ * Applies mixins specified as a __mixins__ property on the given properties
  * object, returning an object containing the mixed in properties.
  */
 function applyMixins(properties) {
-  var mixins = properties.__mixin__
-  if (!is.Array(mixins)) {
+  var mixins = properties.__mixins__
+  if (type(mixins) != 'array') {
     mixins = [mixins]
   }
   var mixedProperties = {}
   for (var i = 0, l = mixins.length; i < l; i++) {
     mixin(mixedProperties, mixins[i])
   }
-  delete properties.__mixin__
-  return object.extend(mixedProperties, properties)
+  delete properties.__mixins__
+  return extend(mixedProperties, properties)
 }
 
 /**
@@ -59,18 +80,18 @@ function inheritFrom(parentConstructor, childConstructor, prototypeProps, constr
   // Base constructors should only have the properties they're defined with
   if (parentConstructor !== Concur) {
     // Inherit the parent's prototype
-    object.inherits(childConstructor, parentConstructor)
+    inherits(childConstructor, parentConstructor)
     childConstructor.__super__ = parentConstructor.prototype
   }
 
   // Add prototype properties - this is why we took a copy of the child
   // constructor reference in extend() - if a .constructor had been passed as a
-  // __mixin__ and overitten prototypeProps.constructor, these properties would
+  // __mixins__ and overitten prototypeProps.constructor, these properties would
   // be getting set on the mixed-in constructor's prototype.
-  object.extend(childConstructor.prototype, prototypeProps)
+  extend(childConstructor.prototype, prototypeProps)
 
   // Add constructor properties
-  object.extend(childConstructor, constructorProps)
+  extend(childConstructor, constructorProps)
 
   return childConstructor
 }
@@ -81,7 +102,7 @@ function inheritFrom(parentConstructor, childConstructor, prototypeProps, constr
 var Concur = module.exports = function() {}
 
 /**
- * Details of a coonstructor's inheritance chain - Concur just facilitates sugar
+ * Details of a constructor's inheritance chain - Concur just facilitates sugar
  * so we don't include it in the initial chain. Arguably, Object.prototype could
  * go here, but it's just not that interesting.
  */
@@ -105,13 +126,15 @@ Concur.extend = function(prototypeProps, constructorProps) {
 
   // Any child constructor passed in should take precedence - grab a reference
   // to it befoer we apply any mixins.
-  var childConstructor = object.get(prototypeProps, 'constructor', null)
+  var childConstructor = (hasOwn.call(prototypeProps, 'constructor')
+                          ? prototypeProps.constructor
+                          : null)
 
   // If any mixins are specified, mix them into the property objects
-  if (object.hasOwn(prototypeProps, '__mixin__')) {
+  if (hasOwn.call(prototypeProps, '__mixins__')) {
     prototypeProps = applyMixins(prototypeProps)
   }
-  if (object.hasOwn(constructorProps, '__mixin__')) {
+  if (hasOwn.call(constructorProps, '__mixins__')) {
     constructorProps = applyMixins(constructorProps)
   }
 
@@ -130,163 +153,5 @@ Concur.extend = function(prototypeProps, constructorProps) {
   return childConstructor
 }
 
-},{"isomorph/is":2,"isomorph/object":3}],2:[function(require,module,exports){
-var toString = Object.prototype.toString
-
-// Type checks
-
-function isArray(o) {
-  return toString.call(o) == '[object Array]'
-}
-
-function isBoolean(o) {
-  return toString.call(o) == '[object Boolean]'
-}
-
-function isDate(o) {
-  return toString.call(o) == '[object Date]'
-}
-
-function isError(o) {
-  return toString.call(o) == '[object Error]'
-}
-
-function isFunction(o) {
-  return toString.call(o) == '[object Function]'
-}
-
-function isNumber(o) {
-  return toString.call(o) == '[object Number]'
-}
-
-function isObject(o) {
-  return toString.call(o) == '[object Object]'
-}
-
-function isRegExp(o) {
-  return toString.call(o) == '[object RegExp]'
-}
-
-function isString(o) {
-  return toString.call(o) == '[object String]'
-}
-
-// Content checks
-
-function isEmpty(o) {
-  for (var prop in o) {
-    return false
-  }
-  return true
-}
-
-module.exports = {
-  Array: isArray
-, Boolean: isBoolean
-, Date: isDate
-, Empty: isEmpty
-, Error: isError
-, Function: isFunction
-, NaN: isNaN
-, Number: isNumber
-, Object: isObject
-, RegExp: isRegExp
-, String: isString
-}
-
-},{}],3:[function(require,module,exports){
-/**
- * Callbound version of Object.prototype.hasOwnProperty(), ready to be called
- * with an object and property name.
- */
-var hasOwn = (function() {
-  var hasOwnProperty = Object.prototype.hasOwnProperty
-  return function(obj, prop) { return hasOwnProperty.call(obj, prop) }
-})()
-
-/**
- * Copies own properties from any given objects to a destination object.
- */
-function extend(dest) {
-  for (var i = 1, l = arguments.length, src; i < l; i++) {
-    src = arguments[i]
-    if (src) {
-      for (var prop in src) {
-        if (hasOwn(src, prop)) {
-          dest[prop] = src[prop]
-        }
-      }
-    }
-  }
-  return dest
-}
-
-/**
- * Makes a constructor inherit another constructor's prototype without
- * having to actually use the constructor.
- */
-function inherits(childConstructor, parentConstructor) {
-  var F = function() {}
-  F.prototype = parentConstructor.prototype
-  childConstructor.prototype = new F()
-  childConstructor.prototype.constructor = childConstructor
-  return childConstructor
-}
-
-/**
- * Creates an Array of [property, value] pairs from an Object.
- */
-function items(obj) {
-  var items = []
-  for (var prop in obj) {
-    if (hasOwn(obj, prop)) {
-      items.push([prop, obj[prop]])
-    }
-  }
-  return items
-}
-
-/**
- * Creates an Object from an Array of [property, value] pairs.
- */
-function fromItems(items) {
-  var obj = {}
-  for (var i = 0, l = items.length, item; i < l; i++) {
-    item = items[i]
-    obj[item[0]] = item[1]
-  }
-  return obj
-}
-
-/**
- * Creates a lookup Object from an Array, coercing each item to a String.
- */
-function lookup(arr) {
-  var obj = {}
-  for (var i = 0, l = arr.length; i < l; i++) {
-    obj[''+arr[i]] = true
-  }
-  return obj
-}
-
-/**
- * If the given object has the given property, returns its value, otherwise
- * returns the given default value.
- */
-function get(obj, prop, defaultValue) {
-  return (hasOwn(obj, prop) ? obj[prop] : defaultValue)
-}
-
-module.exports = {
-  hasOwn: hasOwn
-, extend: extend
-, inherits: inherits
-, items: items
-, fromItems: fromItems
-, lookup: lookup
-, get: get
-}
-
-},{}]},{},[1])
-(1)
+},{}]},{},[1])(1)
 });